instagram_monitor.py
import os
import time
import requests
from telegram import Bot
from telegram.ext import Updater, CommandHandler
from apscheduler.schedulers.background import BackgroundScheduler
from apscheduler.triggers.interval import IntervalTrigger
import pytz
from datetime import datetime
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

# Configuration
TELEGRAM_TOKEN = os.getenv('TELEGRAM_TOKEN', 'YOUR_TELEGRAM_BOT_TOKEN')
CHAT_ID = os.getenv('CHAT_ID', 'YOUR_CHAT_ID')
INSTAGRAM_USERNAME = 'target_instagram_username'  # Replace with the username to monitor

# Initialize bot
bot = Bot(token=TELEGRAM_TOKEN)

def send_message(message):
    """Send a message to the specified chat."""
    try:
        bot.send_message(chat_id=CHAT_ID, text=message)
        print(f"Message sent: {message}")
    except Exception as e:
        print(f"Error sending message: {e}")

def check_instagram_activity():
    """Check for new Instagram activity (placeholder function)."""
    # This is a placeholder. In a real implementation, you would use Instagram's API
    # or web scraping to check for new posts, stories, etc.
    current_time = datetime.now(pytz.timezone('UTC')).strftime('%Y-%m-%d %H:%M:%S')
    message = f"Instagram check at {current_time} - No new activity detected."
    send_message(message)

def start_scheduler():
    """Start the background scheduler."""
    scheduler = BackgroundScheduler(timezone=pytz.timezone('UTC'))
    trigger = IntervalTrigger(minutes=10)  # Check every 10 minutes
    scheduler.add_job(check_instagram_activity, trigger)
    scheduler.start()
    print("Scheduler started. Monitoring Instagram every 10 minutes.")

def main():
    """Main function to run the bot."""
    print("Starting Instagram Monitor Bot...")

    # Start the scheduler
    start_scheduler()

    # Keep the script running
    try:
        while True:
            time.sleep(1)
    except KeyboardInterrupt:
        print("Stopping the bot...")
        scheduler.shutdown()

if __name__ == '__main__':
    main()
