requirements.txt
python-telegram-bot==20.7
pytz==2025.2
python-dotenv==1.0.0
requests==2.32.5
apscheduler==3.11.0
